Machine learning algorithm using camera
This algorithm is the same algorithm used in the previous Arduino but we detect incoming cars/objects using a camera instead of a sensor. We're using the YOLO_v8 algorithm for first detecting if the incoming object is a car or not and if it is then it will use a reference object in the video to measure the speed, distance and acceleration of the incoming car to determine if there will be a potential crash.